Fix save method

+++ building/buildings/models.py
@@ -2,6 +2,7 @@ import os
 
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+from django.core.exceptions import ObjectDoesNotExist
 
 
 class Building(models.Model):
@@ -89,10 +90,19 @@ class ApartmentImage(models.Model):
         using=None,
         update_fields=None,
     ):
-        old_version = ApartmentImage.objects.get(pk=self.pk)
+        try:
+            old_version = ApartmentImage.objects.get(pk=self.pk)
+        except ObjectDoesNotExist:
+            super().save(
+                force_insert=force_insert,
+                force_update=force_update,
+                using=using,
+                update_fields=update_fields,
+            )
+            return 0
         path = old_version.image.path
 
-        if os.path.exists(path):
+        if os.path.exists(path) and self.image != old_version.image:
             os.remove(path)
 
         super().save(
